refactor: move DispatchModal pure functions to dispatchBriefUtils.ts
Fixes the fast-refresh warning — DispatchModal.tsx was exporting both a
React component and plain functions, which breaks Vite HMR. Moves
ShapSignal, parseSignals, buildPdfFilename, buildExportPayload, and
buildCopyMarkdown into a dedicated dispatchBriefUtils.ts file. Updates
DispatchModal.tsx and DispatchModal.test.ts to import from there.
